+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - fixed `ProofTree` handling of getting rules from spec
+- fixed printing of InvalidOperationError message
 
 ## [2.0.0] - 2020-07-16
 ### Added